requirements.txt
flet 
miskibin
matplotlib
pandas
seaborn
pyyaml
pyinstaller